requirements.txt
aiogram
python-dotenv
pytz
timezonefinder
apscheduler
aiosqlite
aiohttp
geopy
dataclasses